Problem: mock query dec doesn't return withdrawtx
Solution: moved TxWithOutputs to chain core and return directly in query responses
